requirements.txt
appnope==0.1.3
asttokens==2.2.1
backcall==0.2.0
cffi==1.15.1
cloudpickle==2.2.1
colorama==0.4.6
comm==0.1.3
contourpy==1.0.7
cycler==0.11.0
debugpy==1.6.7
decorator==5.1.1
executing==1.2.0
filelock==3.10.7
fonttools==4.39.3
gymnasium==0.27.1
-e ../gymnasium-custom
gymnasium-notices==0.0.1
importlib-metadata==6.1.0
importlib-resources==5.12.0
ipykernel==6.22.0
ipython==8.12.0
jax-jumpy==1.0.0
jedi==0.18.2
Jinja2==3.1.2
jupyter_client==8.1.0
jupyter_core==5.3.0
kiwisolver==1.4.4
MarkupSafe==2.1.2
matplotlib==3.7.1
matplotlib-inline==0.1.6
mpmath==1.3.0
nest-asyncio==1.5.6
networkx==3.1
numpy==1.24.2
packaging==23.0
pandas==1.5.3
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.5.0
platformdirs==3.2.0
prompt-toolkit==3.0.38
psutil==5.9.4
ptyprocess==0.7.0
pure-eval==0.2.2
pycparser==2.21
pygame==2.3.0
Pygments==2.14.0
pyparsing==3.0.9
pyreadr==0.4.7
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2022.7.1
pytz-deprecation-shim==0.1.0.post0
pyzmq==25.0.2
six==1.16.0
stack-data==0.6.2
sympy==1.11.1
torch==2.0.0
tornado==6.2
traitlets==5.9.0
typing_extensions==4.5.0
tzdata==2023.3
tzlocal==4.3
wcwidth==0.2.6
zipp==3.15.0
cmake
lit
jupyterlab==3.6.3

